fix(cli): Fix all unit tests
